+++ CHANGELOG.md
@@ -6,6 +6,9 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Upgraded runtime Dart SDK to `2.13.4`
  * Upgraded dependencies.
  * NOTE: enabled activity log UI without experimental flag.
+ * NOTE: added `PackageVersion.isRetracted`.
+         TODO(deferred): make it required, add to integrity checks.
+ * NOTE: added weekly (versioned) periodic task: `backfill-new-fields`.
 
 ## `20210722t120000-all`
  * Bumped runtimeVersion to `2021.07.21`.